Refactor Jira code
