+++ estimage/plugins/jira/importer.py
@@ -55,13 +55,7 @@ class JiraWithRetry(JIRA):
 
 class BareboneImporter:
     def __init__(self, spec):
-        self._cards_by_id = dict()
         self._all_issues_by_name = dict()
-        self._parent_name_to_children_names = dict()
-
-        self._retro_cards = set()
-        self._projective_cards = set()
-        self._all_events = []
 
         try:
             self.jira = JiraWithRetry(spec.server_url, token_auth=spec.token, validate=True)
@@ -70,9 +64,22 @@ class BareboneImporter:
             raise RuntimeError(msg) from exc
 
         self.item_class = spec.item_class
+        self.expand = []
+        self.fields = ["summary"]
 
     def report(self, msg):
         print(msg)
+
+    def _execute_search_query(self, query):
+        items = self.jira.search_issues(query, expand=self.expand, maxResults=0)
+        return items
+
+    def perform_and_process_query(self, query) -> set:
+        results = self._execute_search_query(query)
+        results_by_name = {r.key: r for r in results}
+        self._all_issues_by_name.update(results_by_name)
+        got_names = set(results_by_name.keys())
+        return got_names
 
     def find_card(self, name: str, expand=""):
         card = self.jira.issue(name, expand=expand)
